requirements.txt
attrs==20.3.0
aws-cdk.assets==1.147.0
aws-cdk.aws-acmpca==1.147.0
aws-cdk.aws-apigateway==1.147.0
aws-cdk.aws-applicationautoscaling==1.147.0
aws-cdk.aws-autoscaling==1.147.0
aws-cdk.aws-autoscaling-common==1.147.0
aws-cdk.aws-autoscaling-hooktargets==1.147.0
aws-cdk.aws-certificatemanager==1.147.0
aws-cdk.aws-cloudformation==1.147.0
aws-cdk.aws-cloudfront==1.147.0
aws-cdk.aws-cloudwatch==1.147.0
aws-cdk.aws-codebuild==1.147.0
aws-cdk.aws-codecommit==1.147.0
aws-cdk.aws-codeguruprofiler==1.147.0
aws-cdk.aws-codestarnotifications==1.147.0
aws-cdk.aws-cognito==1.147.0
aws-cdk.aws-dynamodb==1.147.0
aws-cdk.aws-ec2==1.147.0
aws-cdk.aws-ecr==1.147.0
aws-cdk.aws-ecr-assets==1.147.0
aws-cdk.aws-ecs==1.147.0
aws-cdk.aws-efs==1.147.0
aws-cdk.aws-eks==1.147.0
aws-cdk.aws-elasticloadbalancing==1.147.0
aws-cdk.aws-elasticloadbalancingv2==1.147.0
aws-cdk.aws-events==1.147.0
aws-cdk.aws-globalaccelerator==1.147.0
aws-cdk.aws-iam==1.147.0
aws-cdk.aws-kinesis==1.147.0
aws-cdk.aws-kms==1.147.0
aws-cdk.aws-lambda==1.147.0
aws-cdk.aws-logs==1.147.0
aws-cdk.aws-route53==1.147.0
aws-cdk.aws-route53-targets==1.147.0
aws-cdk.aws-s3==1.147.0
aws-cdk.aws-s3-assets==1.147.0
aws-cdk.aws-sagemaker==1.147.0
aws-cdk.aws-sam==1.147.0
aws-cdk.aws-secretsmanager==1.147.0
aws-cdk.aws-servicediscovery==1.147.0
aws-cdk.aws-signer==1.147.0
aws-cdk.aws-sns==1.147.0
aws-cdk.aws-sns-subscriptions==1.147.0
aws-cdk.aws-sqs==1.147.0
aws-cdk.aws-ssm==1.147.0
aws-cdk.aws-stepfunctions==1.147.0
aws-cdk.aws-stepfunctions-tasks==1.147.0
aws-cdk.cloud-assembly-schema==1.147.0
aws-cdk.core==1.147.0
aws-cdk.custom-resources==1.147.0
aws-cdk.cx-api==1.147.0
aws-cdk.lambda-layer-awscli==1.147.0
aws-cdk.lambda-layer-kubectl==1.147.0
aws-cdk.lambda-layer-node-proxy-agent==1.147.0
aws-cdk.region-info==1.147.0
boto3==1.21.12
botocore==1.24.12
cattrs==1.10.0
constructs==3.3.233
dill==0.3.4
google-pasta==0.2.0
importlib-metadata==4.11.2
jmespath==0.10.0
jsii==1.54.0
multiprocess==0.70.12.2
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pathos==0.2.8
pox==0.3.0
ppft==1.6.6.4
protobuf==3.19.4
protobuf3-to-dict==0.1.5
publication==0.0.3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.2
sagemaker==2.77.1
six==1.16.0
smdebug-rulesconfig==1.0.1
typing-extensions==4.1.1
urllib3==1.26.8
zipp==3.7.0
